debug model.py
debugging and
deep ResNets of scaling (s) and translation (t) are now defined in coupling layer class, so it's instantiations will have the same architecture for each layer in RNVP.
Currently it takes a 1d tensor of length 2.
Next step can be to make it take a batch of points.